Rename vuex modules; make smoother login.
observatory.js was renamed to instrument_state.js, to reflect its content being the status of online devices.
device_selection.js was renamed to observatory_configuration, because it stores the instrument configurations in each observatory, as well as the 'active' devices in the UI controls.

Logging in automatically refreshes the observatory configuration so the user doesn't have to do a full page refresh to get the correct configuration to show up in the ctrl page.