Fix bug: current_image not updating with new site.
When switching the active site, the ImageView component would change thumbnails but the active displayed image would not change. Now, the current_image (displayed) will change whenver images are refreshed and there is no current_image, or the site is different between the current_image and the latest refreshed image.