Get site events from config, not aws
The site events calculations that were moved from the site code to the
photonranch-api was never integrated in the site code, and subsequent
updates have rendered the photonranch-api version out of date. So, we
are going to switch back to disaplying the site events from the latest
site config data.