Add object search in telescope command form
Refactor the object name input field into a separate component, including the ability to query simbad and update the telescope coordinate fields.
Add a check in the vuex command_params module to clear the mount_object (object name) if the coordinates change more than a few degrees.
This allows the object name to persist when the user is making small adjustments, but doesn't stick around if the user has changed the target of focus.