Add minor patches for new config formats
Specifically, the new wema configs changed 'latitude' to 'site_latitude' (same for longitude). Also 'name' is now 'site_description'.

Also, change the site operation status to not throw an errorif one of device, weather, or enclosure status is missing, but simply report technical difficulty or offline (the former if at least one of the three is sending).