Store drafted data away so we don't lose it.
When re-rendering the project form, any project data passed to the component is written to the project_params in the store. This effectively nukes any input that has already been made in the creation of a new project, so make sure we save the draft off before showing the project view. Also add user ID to projects table and make it sortable.